🙈 캐시 제거
캐시 제거함
